requirements.txt
GitPython
PyYAML
requests
fedora-messaging
python-dotenv
python-crontab
